you now must be logged in to add product, and if you want to delete you must be owner of product

+++ server/database/routes/products/index.js
@@ -114,7 +114,7 @@ router
 
 // EDIT PRODUCT
 router.post("/:id", upload.single("photos"), (req, res) => {
-  console.log("fuckyou");
+  // console.log("fuckyou");
   const body = req.body;
   const paramsId = req.params.id;
 
@@ -166,30 +166,39 @@ router.post("/:id", upload.single("photos"), (req, res) => {
 router.delete("/:id", (req, res) => {
   console.log("req", req.body);
   console.log("req.body.image_url", req.body.body.image_url);
+  let user_id = Number(req.body.body.user_id);
+  let product_user_id = Number(req.body.body.product.user_id);
   const paramsId = req.params.id;
   const key = req.body.body.image_url.split("/").pop();
+  console.log("user id", user_id);
+  console.log("product_user_id", product_user_id);
 
-  Product.where({
-    id: paramsId
-  })
-    .fetch()
-    .then(product => {
-      new Product({
-        id: paramsId
-      })
-        .destroy()
-        .then(() => {
-          s3.deleteObject(
-            {
-              Bucket: process.env.AWS_BUCKET_NAME,
-              Key: key
-            },
-            function(err, data) {}
-          );
+  if (user_id === product_user_id) {
+    Product.where({
+      id: paramsId
+    })
+      .fetch()
+      .then(product => {
+        new Product({
+          id: paramsId
+        })
+          .destroy()
+          .then(() => {
+            s3.deleteObject(
+              {
+                Bucket: process.env.AWS_BUCKET_NAME,
+                Key: key
+              },
+              function(err, data) {}
+            );
 
-          return res.json("success");
-        });
-    });
+            return res.json("success");
+          });
+      });
+  } else {
+    console.log("you are not the owner of this masterpiece");
+    res.json("you need to be the owner of artwork to delete");
+  }
 });
 
 module.exports = router;
